+++ packages/openapi3/src/lib.ts
@@ -124,6 +124,13 @@ const EmitterOptionsSchema: JSONSchemaType<OpenAPI3EmitterOptions> = {
 export const libDef = {
   name: "@typespec/openapi3",
   diagnostics: {
+    "oneof-union": {
+      severity: "error",
+      messages: {
+        default:
+          "@oneOf decorator can only be used on a union or a model property which type is a union.",
+      },
+    },
     "inconsistent-shared-route-request-visibility": {
       severity: "error",
       messages: {
@@ -139,8 +146,7 @@ export const libDef = {
     "invalid-format": {
       severity: "warning",
       messages: {
-        tsv: "Collection format 'tsv' is not supported in OpenAPI3. Defaulting to type 'string'.",
-        formHeader: paramMessage`Collection format '${"value"}' is not supported in OpenAPI3 headers. Defaulting to type 'string'.`,
+        default: paramMessage`Collection format '${"value"}' is not supported in OpenAPI3 ${"paramType"} parameters. Defaulting to type 'string'.`,
       },
     },
     "resource-namespace": {