+++ packages/openapi3/CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log - @typespec/openapi3
 
-This log was last generated on Tue, 08 Aug 2023 22:32:10 GMT and should not be manually modified.
+This log was last generated on Tue, 19 Sep 2023 19:28:32 GMT and should not be manually modified.
+
+## 0.48.1
+Tue, 19 Sep 2023 19:28:32 GMT
+
+### Patches
+
+- Fix: Correctly generate discriminated union mapping property with multiple visibilities involved
+
+## 0.48.0
+Tue, 12 Sep 2023 21:47:11 GMT
+
+### Updates
+
+- Fixed issue where parameters on a PATCH request marked with visibility "create" did not appear.
+- Changed yaml parser from `js-yaml` to `yaml`
+- Fix: `exclusiveMinimum` and `exclusiveMaximum` properties are booleans in openapi3
+- Use `anyOf` instead of `oneOf` for shared route with different request/response bodies
+- Fix handling of `Record<T>` in `model is` and `model extends`
 
 ## 0.47.0
 Tue, 08 Aug 2023 22:32:10 GMT